feat(core-wave1): extract reminder domain utils into taskify-core and wire PWA
